Reputation now has stars
- You can see them in the shop screen
- I genuinely can't think of anything which could go wrong with this

+++ Assets/Scripts/Player/Reputation.cs
@@ -4,16 +4,16 @@ using UnityEngine;
 
 public class Reputation
 {
-    private int reputation = 0;
+    private float reputation = 0;
 
     public static Reputation instance = new Reputation();
 
-    public static int GetReputation()
+    public static float GetReputation()
     {
         return instance.reputation;
     }
 
-    public static void AddReputation(int add)
+    public static void AddReputation(float add)
     {
         instance.reputation += add;
     }
